Added calculation of candidates on overall nomination

+++ module-server/src/main/java/de/twomartens/wahlrecht/model/SeatResult.java
@@ -5,7 +5,7 @@ import java.util.Map;
 import lombok.Builder;
 
 @Builder
-public record SeatResult(Map<Nomination, Long> seatsPerNomination,
+public record SeatResult(Map<Nomination, Integer> seatsPerNomination,
                          Deque<Double> usedElectionNumbers) {
 
 }
